feat: patch requests status was implemented.

+++ ewm-service/src/main/java/ru/practicum/events/dto/EventRequestStatusUpdateRequestDto.java
@@ -13,6 +13,6 @@ import java.util.List;
 @NoArgsConstructor
 @Builder
 public class EventRequestStatusUpdateRequestDto {
-    private List<Integer> requestIds;
+    private List<Long> requestIds;
     private RequestStatusEnum status;
 }
